pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "gitviz"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "gitpython",
    "click",
    "rich",
]

[project.scripts]
gitviz = "gitviz.cli:main"

[tool.setuptools.packages.find]
where = ["."]